@@ -1,7 +1,7 @@
 {
   "name": "react-html-email",
   "description": "Create elegant HTML email templates using React.",
-  "version": "1.2.0",
+  "version": "3.0.0",
   "keywords": [
     "react",
     "html",
@@ -13,36 +13,38 @@
     "email": "c@chromakode.com"
   },
   "scripts": {
-    "lint": "eslint ./examples ./scripts ./src ./test",
-    "test": "mocha --compilers js:babel-register",
-    "test:cov": "babel-node $(npm bin)/isparta cover node_modules/mocha/bin/_mocha",
+    "lint": "eslint --ext .jsx --ext .js .",
+    "test": "jest",
+    "test:cov": "jest --coverage",
     "compile": "babel -d lib/ src/ && cp -v src/supportMatrix.json lib/",
     "scrape": "babel-node scripts/scrapeStyleSupport.js > src/supportMatrix.json",
-    "prepublish": "npm run compile"
+    "prepare": "npm run compile"
   },
   "main": "lib/index.js",
   "repository": "chromakode/react-html-email",
   "devDependencies": {
-    "cheerio": "~0.19.0",
-    "node-fetch": "~1.3.3",
-    "babel": "~6.3.26",
-    "babel-register": "~6.3.13",
-    "babel-cli": "~6.3.17",
-    "babel-plugin-transform-object-rest-spread": "~6.3.13",
-    "babel-preset-es2015": "~6.3.13",
-    "babel-preset-react": "~6.3.13",
-    "mocha": "~2.3.4",
-    "expect": "~1.13.4",
-    "isparta": "~4.0.0",
-    "eslint": "~1.10.3",
-    "eslint-plugin-react": "~3.11.2",
-    "eslint-config-airbnb": "~2.0.0",
-    "babel-eslint": "~4.1.6",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0"
+    "babel-cli": "~6.26.0",
+    "babel-eslint": "~8.0.0",
+    "babel-jest": "^21.0.2",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
+    "babel-preset-env": "^1.6.0",
+    "babel-preset-react": "^6.24.1",
+    "cheerio": "~1.0.0-rc.2",
+    "eslint": "~4.7.2",
+    "eslint-config-airbnb": "~15.1.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-react": "^7.4.0",
+    "jest": "^21.1.0",
+    "node-fetch": "~1.7.3",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0",
-    "react-dom": "^0.14.0 || ^15.0.0"
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0"
+  },
+  "dependencies": {
+    "prop-types": "^15.5.10"
   }
 }
