@@ -1,18 +1,21 @@
-import React from 'react'
+import PropTypes from 'prop-types'
 import StyleValidator from './StyleValidator'
 
-export const __styleValidator = new StyleValidator()
+export const styleValidator = new StyleValidator()
 
 export function configStyleValidator(config) {
-  __styleValidator.setConfig(config)
+  styleValidator.setConfig(config)
 }
 
 export default {
   style(props, propName, componentName, ...rest) {
-    const objErr = React.PropTypes.object(props, propName, componentName, ...rest)
+    if (props[propName] == null) {
+      return undefined
+    }
+    const objErr = PropTypes.object(props, propName, componentName, ...rest)
     if (objErr) {
       return objErr
     }
-    return __styleValidator.validate(props[propName], componentName)
+    return styleValidator.validate(props[propName], componentName)
   },
 }
