@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import EmailPropTypes from '../PropTypes'
 
 export default function Box({ children, ...props }) {
@@ -31,4 +32,9 @@ Box.defaultProps = {
   border: '0',
   align: 'left',
   valign: 'top',
+  bgcolor: undefined,
+  width: undefined,
+  height: undefined,
+  style: undefined,
+  children: undefined,
 }
