@@ -1,4 +1,5 @@
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import EmailPropTypes from '../PropTypes'
 import Box from './Box'
 import Item from './Item'
@@ -7,21 +8,23 @@ import Item from './Item'
 export default function Email(props) {
   // default nested 600px wide outer table container (see http://templates.mailchimp.com/development/html/)
   return (
-    <html xmlns="http://www.w3.org/1999/xhtml">
+    <html lang={props.lang} xmlns="http://www.w3.org/1999/xhtml">
       <head>
         <meta httpEquiv="Content-Type" content="text/html; charset=utf-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>{props.title}</title>
         {props.headCSS && <style type="text/css">{props.headCSS}</style>}
       </head>
-      <body style={{
-        width: '100%',
-        margin: 0,
-        padding: 0,
-        WebkitTextSizeAdjust: '100%',
-        MsTextSizeAdjust: '100%',
-        ...props.bodyStyle,
-      }}>
+      <body
+        style={{
+          width: '100%',
+          margin: 0,
+          padding: 0,
+          WebkitTextSizeAdjust: '100%',
+          MsTextSizeAdjust: '100%',
+          ...props.bodyStyle,
+        }}
+      >
         <Box width={props.width} bgcolor={props.bgcolor} align="center" cellPadding={props.cellPadding} cellSpacing={props.cellSpacing} style={props.style}>
           {props.children}
         </Box>
@@ -31,21 +34,30 @@ export default function Email(props) {
 }
 
 Email.propTypes = {
+  lang: PropTypes.string,
   title: PropTypes.string.isRequired,
   bgcolor: PropTypes.string,
   cellPadding: PropTypes.number,
   cellSpacing: PropTypes.number,
   style: EmailPropTypes.style,
   headCSS: PropTypes.string,
-  children: PropTypes.node,
   width: PropTypes.string,
   align: PropTypes.oneOf(['left', 'center', 'right']),
   valign: PropTypes.oneOf(['top', 'middle', 'bottom']),
-  bodyStyle: PropTypes.object,
+  bodyStyle: EmailPropTypes.style,
+  children: PropTypes.node,
 }
 
 Email.defaultProps = {
+  lang: 'en',
   width: '600',
   align: 'center',
   valign: 'top',
+  bgcolor: undefined,
+  cellPadding: undefined,
+  cellSpacing: undefined,
+  style: undefined,
+  headCSS: undefined,
+  bodyStyle: undefined,
+  children: undefined,
 }
