@@ -1,7 +1,7 @@
 export default function (props) {
   const dataProps = {}
 
-  Object.keys(props).forEach(key => {
+  Object.keys(props).forEach((key) => {
     if (/^data-/.test(key)) {
       dataProps[key] = props[key]
     }
