@@ -1,17 +1,16 @@
-import PropTypes, { configStyleValidator, __styleValidator } from './PropTypes'
+import PropTypes, { configStyleValidator, styleValidator } from './PropTypes'
 import Box from './components/Box'
 import Email from './components/Email'
 import Image from './components/Image'
 import Item from './components/Item'
 import Span from './components/Span'
 import A from './components/A'
-import injectReactEmailAttributes from './injectReactEmailAttributes'
 import renderEmail from './renderEmail'
 
-const __DEV__ = typeof process !== 'undefined' && process.env && process.env.NODE_ENV !== 'production'
+const DEV = typeof process !== 'undefined' && process.env && process.env.NODE_ENV !== 'production'
 
 configStyleValidator({
-  warn: __DEV__,
+  warn: DEV,
 })
 
 export {
@@ -22,15 +21,13 @@ export {
   Item,
   Span,
   A,
-  injectReactEmailAttributes,
   configStyleValidator,
   renderEmail,
 }
 
 export default {
   PropTypes,
-  injectReactEmailAttributes,
   configStyleValidator,
   renderEmail,
-  __styleValidator,
+  styleValidator,
 }
